feat: implement deleteOne
