+++ src/mongoose/index.ts
@@ -9,6 +9,7 @@ import { GlobalModel } from '../globals/config/types';
 import { find } from './find';
 import { create } from './create';
 import { updateOne } from './updateOne';
+import { deleteOne } from './deleteOne';
 import { findVersions } from './findVersions';
 import { findGlobalVersions } from './findGlobalVersions';
 import type { Payload } from '../index';
@@ -63,5 +64,6 @@ export function mongooseAdapter({ payload, url, connectOptions }: Args): Mongoos
     findGlobalVersions,
     create,
     updateOne,
+    deleteOne,
   };
 }
